+++ lafryhi/routes/users.js
@@ -1,10 +1,12 @@
+const secret = 'fpm@DFG346&ƒgDFG346&ƒgdf#jd]}A{hbsfS43jdf#jd]}A{hbsfS43j';
+var jwt = require('express-jwt');
 var express = require('express');
 var router = express.Router();
 const usersRepo = require('../repositories/users')
 
 router.get('/', async function (req, res, next) {
     const number = req.query.number;
-    if(number === undefined)
+    if (number === undefined)
         res.send(await usersRepo.getAllUsers())
     else
         res.send(await usersRepo.getUsersNumber())
@@ -43,24 +45,49 @@ router.get('/email/:email', async function (req, res, next) {
     res.send(await usersRepo.getUserByEmail(email))
 });
 
-router.delete('/:id', async function (req, res, next) {
-    const id = req.params.id;
-    res.send(await usersRepo.deleteUser(id))
+router.delete('/:id', jwt({algorithms: ['HS256'], secret: secret}), async function (req, res, next) {
+    if (req.user.role === 'admin') {
+        const id = req.params.id;
+        res.send(await usersRepo.deleteUser(id));
+    } else {
+        res.sendStatus(403);
+    }
 });
 
-router.delete('/', async function (req, res, next) {
-    res.send(await usersRepo.deleteAllUsers())
+router.delete('/', jwt({algorithms: ['HS256'], secret: secret}), async function (req, res, next) {
+    if (req.user.role === 'admin') {
+        res.send(await usersRepo.deleteAllUsers());
+    } else {
+        res.sendStatus(403);
+    }
 });
 
-router.post('/', async function (req, res, next) {
-    const user = req.body;
-    res.send(await usersRepo.addUser(user))
+router.post('/', jwt({algorithms: ['HS256'], secret: secret}), async function (req, res, next) {
+    if (req.user.role === 'admin') {
+        const user = req.body;
+        res.send(await usersRepo.addUser(user));
+    } else {
+        res.sendStatus(403);
+    }
 });
 
-router.put('/:id', async function (req, res, next) {
-    const id = req.params.id;
-    const user = req.body;
-    res.send(await usersRepo.updateUser(id, user))
+router.put('/:id', jwt({algorithms: ['HS256'], secret: secret}), async function (req, res, next) {
+    if (req.user.role === 'admin') {
+        const id = req.params.id;
+        const user = req.body;
+        res.send(await usersRepo.updateUser(id, user));
+    } else {
+        res.sendStatus(403);
+    }
+});
+
+router.post('/login', jwt({algorithms: ['HS256'], secret: secret}), async function (req, res, next) {
+    if (req.user.role === 'admin') {
+        const data = req.body;
+        res.send(await usersRepo.login(data));
+    } else {
+        res.sendStatus(403);
+    }
 });
 
 module.exports = router;
